fix(bower): remove version property
Unlike npm, the version property in `bower.json` has no significance (Bower
reads the actual version from Git tags), see [Bower spec][].

This is fortunate for us as versioning is handled by semantic-release (which
creates Git tags for us).

However, if the version property is not removed (and Bower finds a newer Git
tag), a version mismatch warning is shown (which is safe to ignore, but
distracting nonetheless).

Settle any confusion by removing the property altogether.

[bower spec]: https://github.com/bower/bower.json-spec/tree/0b34d2bb0bdd4fa31fa9728f787f49e170e41f99#version